removed console.logs and made slider bar not visible unless issue is selected
