+++ src/actions/db_actions.js
@@ -1,13 +1,11 @@
 import * as dbActions from './types';
 import axios from 'axios';
 
-console.log('the actions in db.actions: ', dbActions);
 //------------------------------------------------------------------------------
 // Since the data is currently not changing we dont need to make a call to the
 // backend every time we select a new contry. For breavity we will only call
 // if the country have not been called yet
 //------------------------------------------------------------------------------
-
 
 //NOTE: For now we will write the logic in the action
 export function getAllData() {
