+++ src/selectors/category_selector.js
@@ -1,7 +1,6 @@
 //NOTE: consider bring in lodash for these selectors
 
 export function issueSelect(data, category) {
-  console.log('args in issueSelect: ', data, category);
   const storage = [];
 
   for (var i = 0; i < data.length; i++) {
@@ -10,12 +9,10 @@ export function issueSelect(data, category) {
     };
   };
 
-  console.log('this is storage in issueSelect: ', storage);
   return storage;
 };
 
 export function colorPopulate(data) {
-  console.log('args in colorPopulate: ', data);
   let range = [];
   let lowrange = undefined;
   let highrange = 0;
@@ -38,6 +35,5 @@ export function colorPopulate(data) {
     }
   }
 
-  console.log('colorPopulate returns: ', [range, lowrange, highrange]);
   return [range, lowrange, highrange];
 }
